Add support for ty, and do some significant internal refactoring

+++ lspinstaller/resolver/resolver.py
@@ -1,3 +1,4 @@
+from lspinstaller.data.types import Source, ArchSpec
 from dataclasses import dataclass
 import enum
 import os
@@ -5,6 +6,7 @@ import stat
 import subprocess
 import shutil
 import platform
+from loguru import logger
 from lspinstaller.config import Config
 from lspinstaller.constants import LSP_HOME
 from lspinstaller.data import sources
@@ -35,27 +37,37 @@ def parse_special(value, data: Data) -> str:
 
         return value
 
-def resolve(package_name, spec, config: Config):
+def ensure_venv():
+    if not os.path.exists(os.path.join(LSP_HOME, "env")):
+        logger.info("venv not set up. Creating it now")
+        logger.warning("This step may fail if you don't have python3-venv installed")
+        subprocess.run(
+            ["python3", "-m", "venv", "env"],
+            cwd=LSP_HOME,
+            check=True,
+        )
+
+def resolve(package_name, spec: Source, config: Config):
     assert spec is not None
     os.makedirs(
         os.path.join(LSP_HOME, "bin"),
         exist_ok=True
     )
-    if "removed" in spec:
-        print(
+    if spec.removed:
+        logger.info(
             f"{package_name} has been removed, and will not be updated or installed"
         )
-    elif "github" in spec:
+    elif spec.github:
         release = get_release_info(
-            spec["github"]["fragment"],
-            spec.get("version_parser", default_version_parser)
+            spec.github.fragment,
+            spec.version_parser
         )
-        binary = spec["binary"]
+        binary = spec.binary
         assert binary is not None, \
             f"misconfigured binary object for {package_name}"
 
-        if "url" in binary:
-            url = binary["url"]
+        if binary.url:
+            url = binary.url
             url = parse_special(
                 url,
                 init_data(
@@ -63,7 +75,7 @@ def resolve(package_name, spec, config: Config):
                 )
             )
         else:
-            pattern = binary["pattern"]
+            pattern = binary.pattern
             assert pattern is not None
             data = init_data(
                 release.tag_name
@@ -76,25 +88,34 @@ def resolve(package_name, spec, config: Config):
                 data,
             )
 
-            if "arch" in spec:
-                arch_spec = spec["arch"]
+            if spec.arch:
+                arch_spec: ArchSpec = spec.arch
                 arch: Arch = resolve_arch()
-                if arch not in arch_spec["supported"][data.os]:
+                if arch not in arch_spec.supported[data.os]:
                     raise RuntimeError(
                         f"{package_name} is not supported on {arch}. "
                         f"Supported: {arch_spec['supported']}"
                     )
 
-                pattern = pattern.replace("${arch}", arch_spec["parser"](arch))
+                pattern = pattern.replace("${arch}", arch_spec.parser(arch))
             url = None
-            for asset in release.assets:
-                if asset.name == pattern:
-                    print(f"Resolved asset to {asset.name}")
-                    url = asset.url
-                    break
+            if binary.pattern_is_url == False:
+                for asset in release.assets:
+                    if asset.name == pattern:
+                        logger.info(f"Resolved asset URL to {asset.name}")
+                        url = asset.url
+                        break
+                else:
+                    logger.error(
+                        f"Failed to resolve GitHub release asset for {package_name}"
+                    )
+                    raise RuntimeError(
+                        f"Failed to resolve GH asset for {package_name}"
+                    )
             else:
-                print(f"Failed to resolve asset for {package_name}")
-                exit(-1)
+                # We've resolved the pattern to its full URL form. Since it's a
+                # URL, it can be used verbatim from this point
+                url = pattern
 
         # Should never throw, but required because url is str | None due to the
         # = None assignment in the previous if statement
@@ -103,11 +124,11 @@ def resolve(package_name, spec, config: Config):
             LSP_HOME,
             url,
             package_name,
-            binary["archive"],
-            binary.get("is_nested", False)
+            binary.archive,
+            binary.is_nested
         )
 
-        for (dest_name, src_path) in binary["link"].items():
+        for (dest_name, src_path) in binary.link.items():
 
             full_src = os.path.join(root, src_path)
             full_dest = os.path.join(
@@ -116,7 +137,7 @@ def resolve(package_name, spec, config: Config):
                 dest_name
             )
             if platform.system() != "Windows":
-                print("Chmoding...")
+                logger.info("Chmoding...")
                 os.chmod(
                     full_src,
                     mode=os.stat(full_src).st_mode | stat.S_IEXEC
@@ -132,20 +153,40 @@ def resolve(package_name, spec, config: Config):
 
         config.update_package(package_name, release.tag_name)
         return release.tag_name
-    elif "npm" in spec:
-        args = ["npm", "install", spec["npm"]["package"]]
-        # TODO: There has to be a shorthand for this
-        if "deps" in spec["npm"]:
-            for dep in spec["npm"]["deps"]:
-                args.append(dep)
+    elif spec.npm:
+        args = [
+            "npm",
+            "install",
+            spec.npm.package
+        ]
+        for dep in spec.npm.deps:
+            args.append(dep)
 
         subprocess.run(
             args,
-            cwd=LSP_HOME
+            cwd=LSP_HOME,
+            check=True,
         )
         return None
+    elif spec.pip:
+        ensure_venv()
+        args = [
+            "./env/bin/python3",
+            "-m",
+            "pip", "install", spec.pip.package
+        ]
+        subprocess.run(
+            args,
+            cwd=LSP_HOME,
+            check=True,
+        )
+        # We need to track the python-managed packages because they cannot be
+        # updated otherwise
+        config.update_package(package_name, "__python_managed__")
+        return "__python_managed__"
 
 def do_update(config: Config):
+    logger.info("Updating npm packages...")
     if os.path.exists(
         os.path.join(
             LSP_HOME,
@@ -160,21 +201,43 @@ def do_update(config: Config):
             cwd=LSP_HOME
         )
 
-    for package in config.packages:
-        dest = os.path.join(
-            LSP_HOME,
-            package
-        )
+    logger.info("Updating other packages...")
+    for [name, package] in config.packages.items():
+        logger.info(f"Updating {name}")
+        try:
+            if "__python_managed__" in package.version:
+                logger.debug("Package is managed by pip; deferring to pip")
+                args = [
+                    "./env/bin/python3",
+                    "-m",
+                    "pip", "install", "--upgrade",
+                    sources[name].pip.package
+                ]
+                subprocess.run(
+                    args,
+                    cwd=LSP_HOME
+                )
+            else:
+                logger.debug("Package is binary; checking for updates")
+                dest = os.path.join(
+                    LSP_HOME,
+                    name
+                )
 
-        # Nuke the existing tree just in case
-        if os.path.exists(dest):
-            shutil.rmtree(
-                dest
-            )
+                # Nuke the existing tree just in case
+                if os.path.exists(dest):
+                    shutil.rmtree(
+                        dest
+                    )
 
-        new_version = resolve(
-            package,
-            sources[package],
-            config
-        )
+                new_version = resolve(
+                    name,
+                    sources[name],
+                    config
+                )
+                assert new_version is not None, \
+                    "Developer error: resolve returned None for binary"
+                package.version = new_version
+        except Exception as e:
+            logger.error(f"Failed to update {name}: {e}")
     config.commit()
